refactor(new-ui): relocate FormikProxy
